Add salt_minion_config_contents config option
Allow a user to define the minion config (which is already YAML)
in the .kitchen.yml file using the `salt_minion_config_contents`
option. The salt_minion_config option already exists, but doesn't
do what it sound like it should (it is a file path that gets joined to
the sandbox path, and the contents of the minion file are put
there... why would anyone want that?).

This allows for a fully customized minion file, for example using
multiple file_roots or an ext_pillar. This is very useful for testing
a whole Salt repo, where you can mount the whole repo into the VM
or container as /srv/salt.

This is hacky right now. Just need to get something working.
For example `formula` and `salt_file_root` are still required in
.kitchen.yml, and must be valid path, even though they aren't used.
Even though they are required, nothing checks to see if they are set,
so an exception is raised only when when the paths are joined or used.

And the salt_minion_config_contents option has not been documented.

Overall kitchen-salt needs a big rework and better docs.